refactor(LoginCard): redesign page
