+++ src/components/icons/HelpCircleOutline.vue
@@ -1,0 +1,13 @@
+<template>
+  <v-icon v-bind="$attrs" v-on="$listeners">{{ svgPath }}</v-icon>
+</template>
+
+<script>
+import { mdiHelpCircleOutline } from '@mdi/js';
+
+export default {
+  data: () => ({
+    svgPath: mdiHelpCircleOutline,
+  }),
+};
+</script>
