+++ src/components/icons/LockOutlineIcon.vue
@@ -1,0 +1,13 @@
+<template>
+  <v-icon v-bind="$attrs" v-on="$listeners">{{ svgPath }}</v-icon>
+</template>
+
+<script>
+import { mdiLockOutline } from '@mdi/js';
+
+export default {
+  data: () => ({
+    svgPath: mdiLockOutline,
+  }),
+};
+</script>
