+++ src/styles/app.scss
@@ -1,0 +1,3 @@
+.vh-100 {
+  height: 100vh !important;
+}
